# services/api/docs/swagger/swagger.yaml
components:
...
  schemas:
...
      GetResult_any.any.any_: { properties: {}, type: object },
      runtime.Types.DefaultSelection_CountryPayload_:
        {
          allOf:
            [
              { $ref: '#/components/schemas/GetResult_any.any.any_' },
              { properties: {}, type: object },
            ],
        },
      Country:
        {
          $ref: '#/components/schemas/runtime.Types.DefaultSelection_CountryPayload_',
          description: 'Model Country',
        },
...
  securitySchemes: {}
info:
  title: '@world/api'
  version: 0.0.1
  description: 'World API'
  license: { name: 'GNU GPL V3.0' }
  contact: { name: 'hieudoanm ', email: hieumdoan@gmail.com }
openapi: 3.0.0
paths:
...
  /countries:
    {
      get:
        {
          operationId: GetCountries,
          responses:
            {
              '200':
                {
                  description: '',
                  content:
                    {
                      application/json:
                        {
                          schema:
                            {
                              items:
                                {
                                  properties:
                                    {
                                      updatedAt:
                                        { type: string, format: date-time },
                                      createdAt:
                                        { type: string, format: date-time },
                                      timezones:
                                        {
                                          items: { type: string },
                                          type: array,
                                        },
                                      googleMaps: { type: string },
                                      density: { type: number, format: double },
                                      population:
                                        { type: number, format: double },
                                      area: { type: number, format: double },
                                      flagSVG: { type: string },
                                      flagPNG: { type: string },
                                      flag: { type: string },
                                      borders:
                                        {
                                          items: { type: string },
                                          type: array,
                                        },
                                      continents:
                                        {
                                          items: { type: string },
                                          type: array,
                                        },
                                      alternativeSpellings:
                                        {
                                          items: { type: string },
                                          type: array,
                                        },
                                      capital:
                                        {
                                          items: { type: string },
                                          type: array,
                                        },
                                      topLevelDomains:
                                        {
                                          items: { type: string },
                                          type: array,
                                        },
                                      startOfWeek: { type: string },
                                      unMember: { type: boolean },
                                      independent: { type: boolean },
                                      status: { type: string },
                                      fifa: { type: string },
                                      cioc: { type: string },
                                      ccn3: { type: string },
                                      cca3: { type: string },
                                      cca2: { type: string },
                                      officialName: { type: string },
                                      commonName: { type: string },
                                      code: { type: string },
                                      longitude:
                                        { type: number, format: double },
                                      latitude:
                                        { type: number, format: double },
                                      subregion: { type: string },
                                      region: { type: string },
                                      cities:
                                        {
                                          items:
                                            {
                                              properties:
                                                {
                                                  timezone:
                                                    {
                                                      type: number,
                                                      format: double,
                                                    },
                                                  longitude:
                                                    {
                                                      type: number,
                                                      format: double,
                                                    },
                                                  latitude:
                                                    {
                                                      type: number,
                                                      format: double,
                                                    },
                                                  city: { type: string },
                                                  state: { type: string },
                                                },
                                              required:
                                                [
                                                  timezone,
                                                  longitude,
                                                  latitude,
                                                  city,
                                                  state,
                                                ],
                                              type: object,
                                            },
                                          type: array,
                                        },
                                      googleTrends:
                                        {
                                          items: { type: string },
                                          type: array,
                                        },
                                      organizations:
                                        {
                                          items:
                                            {
                                              allOf:
                                                [
                                                  {
                                                    $ref: '#/components/schemas/GetResult_any.any.any_',
                                                  },
                                                  {
                                                    properties: {},
                                                    type: object,
                                                  },
                                                ],
                                            },
                                          type: array,
                                        },
                                      languages:
                                        {
                                          items:
                                            {
                                              allOf:
                                                [
                                                  {
                                                    $ref: '#/components/schemas/GetResult_any.any.any_',
                                                  },
                                                  {
                                                    properties: {},
                                                    type: object,
                                                  },
                                                ],
                                            },
                                          type: array,
                                        },
                                      currencies:
                                        {
                                          items:
                                            {
                                              allOf:
                                                [
                                                  {
                                                    $ref: '#/components/schemas/GetResult_any.any.any_',
                                                  },
                                                  {
                                                    properties: {},
                                                    type: object,
                                                  },
                                                ],
                                            },
                                          type: array,
                                        },
                                    },
                                  required:
                                    [
                                      updatedAt,
                                      createdAt,
                                      timezones,
                                      googleMaps,
                                      density,
                                      population,
                                      area,
                                      flagSVG,
                                      flagPNG,
                                      flag,
                                      borders,
                                      continents,
                                      alternativeSpellings,
                                      capital,
                                      topLevelDomains,
                                      startOfWeek,
                                      unMember,
                                      independent,
                                      status,
                                      fifa,
                                      cioc,
                                      ccn3,
                                      cca3,
                                      cca2,
                                      officialName,
                                      commonName,
                                      code,
                                      longitude,
                                      latitude,
                                      subregion,
                                      region,
                                      cities,
                                      googleTrends,
                                      organizations,
                                      languages,
                                      currencies,
                                    ],
                                  type: object,
                                },
                              type: array,
                            },
                        },
                    },
                },
            },
          tags: [Countries],
          security: [],
          parameters:
            [
              {
                in: query,
                name: codes,
                required: false,
                schema: { default: '', type: string },
              },
              {
                in: query,
                name: timezone,
                required: false,
                schema: { default: '', type: string },
              },
            ],
        },
    }
  '/countries/{code}':
    {
      get:
        {
          operationId: GetCountry,
          responses:
            {
              '200':
                {
                  description: '',
                  content:
                    {
                      application/json:
                        { schema: { $ref: '#/components/schemas/Country' } },
                    },
                },
            },
          tags: [Countries],
          security: [],
          parameters:
            [
              {
                in: path,
                name: code,
                required: true,
                schema: { type: string },
              },
            ],
        },
    }
  /countries/search:
    {
      post:
        {
          operationId: SearchCountries,
          responses:
            {
              '200':
                {
                  description: '',
                  content:
                    {
                      application/json:
                        {
                          schema:
                            {
                              properties:
                                {
                                  countries:
                                    {
                                      items:
                                        {
                                          properties:
                                            {
                                              updatedAt:
                                                {
                                                  type: string,
                                                  format: date-time,
                                                },
                                              createdAt:
                                                {
                                                  type: string,
                                                  format: date-time,
                                                },
                                              timezones:
                                                {
                                                  items: { type: string },
                                                  type: array,
                                                },
                                              googleMaps: { type: string },
                                              density:
                                                {
                                                  type: number,
                                                  format: double,
                                                },
                                              population:
                                                {
                                                  type: number,
                                                  format: double,
                                                },
                                              area:
                                                {
                                                  type: number,
                                                  format: double,
                                                },
                                              flagSVG: { type: string },
                                              flagPNG: { type: string },
                                              flag: { type: string },
                                              borders:
                                                {
                                                  items: { type: string },
                                                  type: array,
                                                },
                                              continents:
                                                {
                                                  items: { type: string },
                                                  type: array,
                                                },
                                              alternativeSpellings:
                                                {
                                                  items: { type: string },
                                                  type: array,
                                                },
                                              capital:
                                                {
                                                  items: { type: string },
                                                  type: array,
                                                },
                                              topLevelDomains:
                                                {
                                                  items: { type: string },
                                                  type: array,
                                                },
                                              startOfWeek: { type: string },
                                              unMember: { type: boolean },
                                              independent: { type: boolean },
                                              status: { type: string },
                                              fifa: { type: string },
                                              cioc: { type: string },
                                              ccn3: { type: string },
                                              cca3: { type: string },
                                              cca2: { type: string },
                                              officialName: { type: string },
                                              commonName: { type: string },
                                              code: { type: string },
                                              longitude:
                                                {
                                                  type: number,
                                                  format: double,
                                                },
                                              latitude:
                                                {
                                                  type: number,
                                                  format: double,
                                                },
                                              subregion: { type: string },
                                              region: { type: string },
                                              cities:
                                                {
                                                  items:
                                                    {
                                                      properties:
                                                        {
                                                          timezone:
                                                            {
                                                              type: number,
                                                              format: double,
                                                            },
                                                          longitude:
                                                            {
                                                              type: number,
                                                              format: double,
                                                            },
                                                          latitude:
                                                            {
                                                              type: number,
                                                              format: double,
                                                            },
                                                          city:
                                                            { type: string },
                                                          state:
                                                            { type: string },
                                                        },
                                                      required:
                                                        [
                                                          timezone,
                                                          longitude,
                                                          latitude,
                                                          city,
                                                          state,
                                                        ],
                                                      type: object,
                                                    },
                                                  type: array,
                                                },
                                              googleTrends:
                                                {
                                                  items: { type: string },
                                                  type: array,
                                                },
                                              organizations:
                                                {
                                                  items:
                                                    {
                                                      allOf:
                                                        [
                                                          {
                                                            $ref: '#/components/schemas/GetResult_any.any.any_',
                                                          },
                                                          {
                                                            properties: {},
                                                            type: object,
                                                          },
                                                        ],
                                                    },
                                                  type: array,
                                                },
                                              languages:
                                                {
                                                  items:
                                                    {
                                                      allOf:
                                                        [
                                                          {
                                                            $ref: '#/components/schemas/GetResult_any.any.any_',
                                                          },
                                                          {
                                                            properties: {},
                                                            type: object,
                                                          },
                                                        ],
                                                    },
                                                  type: array,
                                                },
                                              currencies:
                                                {
                                                  items:
                                                    {
                                                      allOf:
                                                        [
                                                          {
                                                            $ref: '#/components/schemas/GetResult_any.any.any_',
                                                          },
                                                          {
                                                            properties: {},
                                                            type: object,
                                                          },
                                                        ],
                                                    },
                                                  type: array,
                                                },
                                            },
                                          required:
                                            [
                                              updatedAt,
                                              createdAt,
                                              timezones,
                                              googleMaps,
                                              density,
                                              population,
                                              area,
                                              flagSVG,
                                              flagPNG,
                                              flag,
                                              borders,
                                              continents,
                                              alternativeSpellings,
                                              capital,
                                              topLevelDomains,
                                              startOfWeek,
                                              unMember,
                                              independent,
                                              status,
                                              fifa,
                                              cioc,
                                              ccn3,
                                              cca3,
                                              cca2,
                                              officialName,
                                              commonName,
                                              code,
                                              longitude,
                                              latitude,
                                              subregion,
                                              region,
                                              cities,
                                              googleTrends,
                                              organizations,
                                              languages,
                                              currencies,
                                            ],
                                          type: object,
                                        },
                                      type: array,
                                    },
                                  total: { type: number, format: double },
                                },
                              required: [countries, total],
                              type: object,
                            },
                        },
                    },
                },
            },
          tags: [Countries],
          security: [],
          parameters: [],
          requestBody:
            {
              required: true,
              content:
                {
                  application/json:
                    {
                      schema:
                        {
                          properties: { query: { type: string } },
                          required: [query],
                          type: object,
                        },
                    },
                },
            },
        },
    }
...
servers:
  - { url: 'http://localhost:8080/api', description: Local }
  - { url: 'http://dashboardworldapi.vercel.app/api', description: Vercel }